requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==2.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2022.1
requests==2.26.0
requests-oauthlib==1.3.1
self==2020.12.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
testfixtures==6.18.5
toml==0.10.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
zipp==3.8.1
